requirements.txt
# Core requirements
streamlit>=1.24.0
speedtest-cli>=2.1.3
plotly>=5.13.0
pandas>=1.5.3
pillow>=9.5.0

# Advanced network testing
ping3>=4.0.0
scapy>=2.5.0
requests>=2.31.0
schedule>=1.2.0
numpy>=1.24.3

# Optional: For improved performance
python-dateutil>=2.8.2
pytz>=2023.3